support reuse and do not reuse connection for bolt
